+++ app/portainer/views/groups/groupsController.js
@@ -1,5 +1,7 @@
 import angular from 'angular';
 import _ from 'lodash-es';
+import { confirmDestructive } from '@@/modals/confirm';
+import { buildConfirmButton } from '@@/modals/utils';
 
 angular.module('portainer.app').controller('GroupsController', GroupsController);
 
@@ -11,14 +13,24 @@ function GroupsController($scope, $state, $async, GroupService, Notifications) {
   }
 
   async function removeActionAsync(selectedItems) {
+    const confirmed = await confirmDestructive({
+      title: '确定吗?',
+      message: '确定要删除所选环境组吗?',
+      confirmButton: buildConfirmButton('删除', 'danger'),
+    });
+
+    if (!confirmed) {
+      return;
+    }
+
     for (let group of selectedItems) {
       try {
         await GroupService.deleteGroup(group.Id);
 
-        Notifications.success('成功删除环境组', group.Name);
+        Notifications.success('环境组已成功删除', group.Name);
         _.remove($scope.groups, group);
       } catch (err) {
-        Notifications.error('失败', err, '无法删除群组');
+        Notifications.error('Failure', err, '无法删除分组');
       }
     }
 
@@ -31,7 +43,7 @@ function GroupsController($scope, $state, $async, GroupService, Notifications) {
         $scope.groups = data;
       })
       .catch(function error(err) {
-        Notifications.error('失败', err, '无法检索环境组');
+        Notifications.error('Failure', err, '无法检索环境组');
         $scope.groups = [];
       });
   }
